# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cashbee Coding Exam Api Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Retrieve current user posts",
                "operationId": "f11765a8d330e514170cca0b0232cb67",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Media type expected by the client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current user posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Velit id tempore autem quasi qui."
                                            },
                                            "banner_image_url": {
                                                "type": "string",
                                                "example": "https://via.placeholder.com/800x400.png/0022ff?text=business+nostrum"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "Officiis qui eveniet nesciunt reprehenderit natus dicta molestiae voluptatum."
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-15 04:44:06"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "PUBLISHED"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-15 04:44:06"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-15 04:44:06"
                                            },
                                            "deleted_at": {
                                                "type": "string",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": {
                    "sanctum": {
                        "bearerAuth": []
                    }
                }
            }
        },
        "/api/posts/all": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Retrieve all posts (admin only)",
                "operationId": "e965d2b5121ae5ac5e24c8d7c4bce43b",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for user authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer JWT_TOKEN_HERE"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Media type expected by the client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Velit id tempore autem quasi qui."
                                            },
                                            "banner_image_url": {
                                                "type": "string",
                                                "example": "https://via.placeholder.com/800x400.png/0022ff?text=business+nostrum"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "Officiis qui eveniet nesciunt reprehenderit natus dicta molestiae voluptatum."
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-15 04:44:06"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "PUBLISHED"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-15 04:44:06"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-15 04:44:06"
                                            },
                                            "deleted_at": {
                                                "type": "string",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorize - User does not exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The following user does not exists."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not have the required role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "You do not have the required role to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/create": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post for the current user",
                "operationId": "e23b6c439c19d51685577415f72d661c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for user authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer JWT_TOKEN_HERE"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Media type expected by the client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My New Post"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "This is the content of my post."
                                    },
                                    "banner_image_url": {
                                        "type": "string",
                                        "example": "https://via.placeholder.com/800x400.png/0022ff?text=business+nostrum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    }
                }
            }
        },
        "/api/posts/{post}/archive": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Archive a post by ID",
                "operationId": "8df2566cf1c794e8a3a8bce1a95ef40c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for user authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer JWT_TOKEN_HERE"
                        }
                    },
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the post to be archived",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Media type expected by the client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post archived successfully"
                    }
                }
            }
        },
        "/api/posts/{post}/restore": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Restore an archived post by ID",
                "operationId": "cb6fcd0bbde4d96e92e733c903f4d498",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for user authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer JWT_TOKEN_HERE"
                        }
                    },
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the post to be restored",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Media type expected by the client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post restored successfully"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "in": "header",
                "bearerAuth": "bearerAuth",
                "bearerFormat": "JWT"
            },
            "sanctum": {
                "bearerAuth": []
            }
        }
    }
}